requirements.txt
streamlit
opencv-python
mediapipe
pandas
numpy
streamlit-lottie
requests
sympy
google-generativeai